A correlation analysis
I wanted to see if some people (Like me) could have rushed through the experiment. Therefore I tried to correlate the result and the time of the test. (And inspired by the complexity paper ) Found that we could take two clusters, one in which the time taken correlates slightly with the result, and one in which there is less correlation. Hope this helps